+++ lib/ua_inspector/config.ex
@@ -24,9 +24,7 @@ defmodule UAInspector.Config do
   def get(key, default \\ nil)
 
   def get(key, default) when is_atom(key) do
-    :ua_inspector
-    |> Application.get_env(key, default)
-    |> maybe_fetch_system()
+    Application.get_env(:ua_inspector, key, default)
   end
 
   def get(keys, default) when is_list(keys) do
@@ -34,7 +32,6 @@ defmodule UAInspector.Config do
     |> Application.get_all_env()
     |> Kernel.get_in(keys)
     |> maybe_use_default(default)
-    |> maybe_fetch_system()
   end
 
   @doc """
@@ -96,33 +93,6 @@ defmodule UAInspector.Config do
       ""
     end
   end
-
-  defp log_system_config_deprecation do
-    Logger.info(fn ->
-      "Accessing the system environment for configuration via" <>
-        " {:system, \"var\"} has been deprecated. Please switch" <>
-        " to an initializer function to avoid future problems."
-    end)
-  end
-
-  defp maybe_fetch_system(config) when is_list(config) do
-    Enum.map(config, fn
-      {k, v} -> {k, maybe_fetch_system(v)}
-      other -> other
-    end)
-  end
-
-  defp maybe_fetch_system({:system, var, default}) do
-    _ = log_system_config_deprecation()
-    System.get_env(var) || default
-  end
-
-  defp maybe_fetch_system({:system, var}) do
-    _ = log_system_config_deprecation()
-    System.get_env(var)
-  end
-
-  defp maybe_fetch_system(config), do: config
 
   defp maybe_use_default(nil, default), do: default
   defp maybe_use_default(config, _), do: config