+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Backwards incompatible changes
     - Minimum required elixir version is now `~> 1.5`
+    - Support for `{:system, var}` configuration has been removed
 
 ## v0.20.0 (2019-03-10)